couple of refactor:   - delete test README.md   - move `hospital_list.yml` and `hospital_list.json` to `resource/API_resource` to make the root looks cleaner   - deletes scr for the code there is out of date   - rename `code_resource_python` to `python_resource`, to make the resource folder cleaner
