+++ resource/python_resource/json_writer.py
@@ -1,4 +1,5 @@
 import json
+import os
 
 try:
     import helpers
@@ -9,8 +10,9 @@ except ImportError:
 
 
 def write(info_dict):
+    filename = os.path.join(os.getcwd(), 'resource', 'API_resource', 'hospital_list.json')
     output = json.dumps(info_dict, ensure_ascii=False)
-    helpers.write_file(filename='hospital_list.json', content=output)
+    helpers.write_file(filename=filename, content=output)
 
 
 if __name__ == '__main__':
