+++ resource/python_resource/readme_reader.py
