+++ resource/python_resource/yaml_reader.py
