--- src/radical.html
@@ -1,89 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf-8">
-<style>
-
-.node circle {
-  fill: #999;
-}
-
-.node text {
-  font: 6.3px sans-serif;
-}
-
-.node--internal circle {
-  fill: #555;
-}
-
-.node--internal text {
-  text-shadow: 0 1px 0 #fff, 0 -1px 0 #fff, 1px 0 0 #fff, -1px 0 0 #fff;
-}
-
-.link {
-  fill: none;
-  stroke: #555;
-  stroke-opacity: 0.5;
-  stroke-width: 1.5px;
-}
-
-</style>
-<body>
-<script charset="utf-8" src="https://d3js.org/d3.v4.0.0-alpha.29.min.js"></script>
-<script charset="utf-8" >
-
-var width = 1500,
-    height = 1350,
-    radius = 700;
-
-var stratify = d3.stratify()
-    .parentId(function(d) { return d.id.substring(0, d.id.lastIndexOf(".")); });
-
-var cluster = d3.cluster()
-    .size([360, radius - 120]);
-
-var svg = d3.select("body").append("svg")
-    .attr("width", width)
-    .attr("height", height)
-  .append("g")
-    .attr("transform", "translate(" + width / 2 + "," + (height / 2 + 20) + ")");
-
-d3.csv("result.csv", function(error, data) {
-  if (error) throw error;
-  var root = stratify(data)
-      .sort(function(a, b) { return a.height - b.height || a.id.localeCompare(b.id); });
-
-  cluster(root);
-
-  var link = svg.selectAll(".link")
-      .data(root.descendants().slice(1))
-    .enter().append("path")
-      .attr("class", "link")
-      .attr("d", function(d) {
-        return "M" + project(d.x, d.y)
-            + "C" + project(d.x, (d.y + d.parent.y) / 2)
-            + " " + project(d.parent.x, (d.y + d.parent.y) / 2)
-            + " " + project(d.parent.x, d.parent.y);
-      });
-
-  var node = svg.selectAll(".node")
-      .data(root.descendants())
-    .enter().append("g")
-      .attr("class", function(d) { return "node" + (d.children ? " node--internal" : " node--leaf"); })
-      .attr("transform", function(d) { return "translate(" + project(d.x, d.y) + ")"; });
-
-  node.append("circle")
-      .attr("r", 2.0);
-
-  node.append("text")
-      .attr("dy", ".31em")
-      .attr("x", function(d) { return d.x < 180 === !d.children ? 6 : -6; })
-      .style("text-anchor", function(d) { return d.x < 180 === !d.children ? "start" : "end"; })
-      .attr("transform", function(d) { return "rotate(" + (d.x < 180 ? d.x - 90 : d.x + 90) + ")"; })
-      .text(function(d) { return d.id.substring(d.id.lastIndexOf(".") + 1); });
-});
-
-function project(x, y) {
-  var angle = (x - 90) / 180 * Math.PI, radius = y;
-  return [radius * Math.cos(angle), radius * Math.sin(angle)];
-}
-
-</script>
